Update LICENSE
This repository is licensed under the MIT License, a permissive open-source license that allows others to freely use, modify, and distribute the code, even for commercial purposes, as long as proper credit is given. This license promotes open collaboration while limiting liability and warranty claims. By including this license, we encourage the community to build upon this work for scientific, educational, or practical applications.